+++ zanata-client-commands/src/main/java/org/zanata/client/commands/glossary/delete/GlossaryDeleteOptions.java
@@ -3,7 +3,7 @@ package org.zanata.client.commands.glossary.delete;
 import org.zanata.client.commands.ConfigurableProjectOptions;
 
 public interface GlossaryDeleteOptions extends ConfigurableProjectOptions {
-    public String getlang();
+    public String getId();
 
     public boolean getAllGlossary();
 }