+++ docs/release-notes.md
@@ -1,6 +1,7 @@
 ## 3.8.0
 
 ##### Changes
+* [ZNTA-354](https://zanata.atlassian.net/browse/ZNTA-354) - Improve Zanata client installation documentation and workflow
 * [ZNTA-664](https://zanata.atlassian.net/browse/ZNTA-664) - Change client push command not to run copy trans by default
 
 -----------------------